Laying out the basics for this utility program
This program goal is to allow a series of developers to create
diagnostic plugins and deploy them in a frame work to automatically
collect logs and other files. This will be accomplished by implementing
this program, the goal will be to create a basic Web app to interact
with these components to allow the information to get where it needs to
be.